mdm_server/requirements.txt
flask>=2.3
pyotp>=2.9
PyJWT>=2.8
requests>=2.31
